+++ lib/open_budget_web/router.ex
@@ -16,13 +16,13 @@ defmodule OpenBudgetWeb.Router do
   scope "/api", OpenBudgetWeb do
     pipe_through :api
 
-    resources "/accounts", AccountController, except: [:new, :edit]
     resources "/users", UserController, only: [:create]
   end
 
   scope "/api", OpenBudgetWeb do
     pipe_through :api_auth
 
+    resources "/accounts", AccountController, except: [:new, :edit]
     resources "/budgets", BudgetController, except: [:new, :edit] do
       post "/switch", BudgetController, :switch, as: :switch
     end